claude-portal

+++ user/backend/src/main/java/com/ebrain/user/dto/response/FreeBoardDto.java
@@ -6,14 +6,14 @@ import java.time.LocalDateTime;
 @Data
 public class FreeBoardDto {
     private Long boardId;
-    private Integer categoryId;
+    private Long categoryId;
     private String categoryName;
     private String authorType;
     private String authorId;
     private String authorName;
     private String title;
     private String content;
-    private Integer views;
+    private Long views;
     private Boolean isDeleted;
     private LocalDateTime createdAt;
     private LocalDateTime editedAt;
