+++ user/backend/src/main/java/com/ebrain/user/dto/response/NoticeBoardDto.java
@@ -6,12 +6,12 @@ import java.time.LocalDateTime;
 @Data
 public class NoticeBoardDto {
     private Long boardId;
-    private Integer categoryId;
+    private Long categoryId;
     private String categoryName;
     private String title;
     private String content;
     private String authorName;
-    private Integer views;
+    private Long views;
     private Boolean isFixed;
     private LocalDateTime createdAt;
     private LocalDateTime editedAt;
